+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 - Restored device labels on split-debugger run tabs in IntelliJ IDEA 2025.3+. (#8908)
+- Debug window opening for run app in 2026.1. (#8985)
 
 ## 92.0.0